+++ CHANGES.md
@@ -13,6 +13,7 @@
 - #3426 People with orgUnit constraint permission should only be able to perform read operation on datasets not assigned to an orgUnit
 - #3424 Set the default order by field for APIs
 - #3427 Dataset Editor: offer more details via error message box when encounter insufficient permission error
+- #3425 offer a button on dataset page to take the user back to data management tab
 
 ## v2.1.1